requirements.txt
PyQt5==5.14.2
python-vlc==3.0.9113 